changes to serial processing of commands from the driver
Because the port contained spurious characters when AZ was sent the first time from the driver it caused clients such as POTH to lock. So instead of  using equalsignorecase, the code now detects the presence of the known characters such as AZ, OS, CS and SS by using INDEXOF to detect the pair in a string. Needs test.