Significantly different code
This code is designed to send AZ, SS, OS and CS and for the two which expect responses (AZ and SS), the responses are validated before sending to the driver. So for AZ, a check is made that a double in the  range 0> value <361 noting that 0 is a valid value , but unfortunately the arduino toDouble function returns zero if there's a conversion error perhaps anothe commit will sort that potential bug. For SS validation consists of checking is OPEN# or CLOSED# is returned. If the validations are not successful retries are made until they are.