NEW COMMS CHECK AND RESTRUCTURE
Two routines created to house the AZ actions and SS actions - no changes to that code, just placed in void AZactions and Void SSactions.
The comms check now writes radio to AZ and SS and diplys the results on the Master radio LCD. Lines 0 and 1 are for Az related info, lines 2 and 3 are for shutter related info.