This has been gvdljoh shredded - the nrf radion was horrific in its coding complecity.
So this version is I think about one tenth of the code (!) and has been tested with a paired bluetooth device on another machine. The commands it sends via Bluetooth are OS, CS, SS and ES. All have been tested on the external bluetooth machine and they work as expected. Receipts also work. So essentially this is complete. It took months to write the NRF radio.